implement pa3 exception syscall fs and devices

+++ nemu/include/memory/mmu.h
@@ -76,7 +76,7 @@ typedef union GateDescriptor {
     uint32_t present          : 1;
     uint32_t offset_31_16     : 16;
   };
-  uint32_t val;
+  uint64_t val;
 } GateDesc;
 
 #endif
